saved code from competition to Comp1
