+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AutoPrograms/Auto_LongRight.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.AutoPrograms;
 
 import com.arcrobotics.ftclib.drivebase.MecanumDrive;
 import com.arcrobotics.ftclib.hardware.motors.Motor;
@@ -8,26 +8,16 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.IMU;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
-
-import com.arcrobotics.ftclib.gamepad.GamepadEx;
-import com.arcrobotics.ftclib.hardware.RevIMU;
-
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.robotcore.external.navigation.YawPitchRollAngles;
-
-@Autonomous(name = "Short_Right",group = "")
-public class Auto_ShortRight extends LinearOpMode {
+@Autonomous(name = "Long_Right",group = "")
+public class Auto_LongRight extends LinearOpMode {
     private final ElapsedTime runtime = new ElapsedTime();
     private Motor leftFrontDrive = null;
     private Motor rightFrontDrive = null;
     private Motor leftBackDrive = null;
     private Motor rightBackDrive = null;
-    private Servo finger;
 
     private Motor arm1 = null;
+    private Servo finger;
     MecanumDrive drivebase = null;
     private IMU imu;// BHI260AP imu on this hub
     private boolean test = false;
@@ -40,17 +30,14 @@ public class Auto_ShortRight extends LinearOpMode {
         leftBackDrive = new Motor(hardwareMap, "left_back_drive");
         rightBackDrive = new Motor(hardwareMap, "right_back_drive");
 
-         imu = hardwareMap.get(IMU.class, "imu");// need to use IMU in expansion hub, not control hub
+        IMU imu = hardwareMap.get(IMU.class, "imu");// need to use IMU in expansion hub, not control hub
         // need to confirm orientation of the HUB so that IMU directions are correct
 
         RevHubOrientationOnRobot.LogoFacingDirection logoDirection = RevHubOrientationOnRobot.LogoFacingDirection.UP;
         RevHubOrientationOnRobot.UsbFacingDirection  usbDirection  = RevHubOrientationOnRobot.UsbFacingDirection.RIGHT;
 
         orientationOnRobot = new RevHubOrientationOnRobot(logoDirection, usbDirection);
-        imu.initialize(new IMU.Parameters(orientationOnRobot));
 
-        YawPitchRollAngles orientation = imu.getRobotYawPitchRollAngles();
-        double heading = orientation.getYaw(AngleUnit.DEGREES);
         drivebase = new MecanumDrive(
                 leftFrontDrive,
                 rightFrontDrive,
@@ -61,26 +48,44 @@ public class Auto_ShortRight extends LinearOpMode {
         waitForStart();
         imu.resetYaw();
         // Step through each leg of the path, ensuring that the Auto mode has not been stopped along the way
-
-        // Step 1:  strafe right for X seconds:
+        //step 1 : move forward to clear frame
+        // Step 2:  strafe right for X seconds:
         // Replace with encoder measured distance if needed
 
+        try {
+            Thread.sleep(10000);
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
         runtime.reset(); // reset timer
-        while (opModeIsActive() && (runtime.seconds() < 1.5)) {
+        //move forward 2 inches
+
+        while (opModeIsActive() && (runtime.seconds() < 0.04)){
+            // tell ftclib its inputs  strafeSpeed,forwardSpeed,turn,heading
+            drivebase.driveFieldCentric(
+                    0,
+                    0.5,
+                    0,
+                    0
+            );
+            telemetry.addData("Path", "Leg 1: %4.1f S Elapsed", runtime.seconds());
+            telemetry.update();
+        }
+
+        while (opModeIsActive() && (runtime.seconds() < 3.0)) {
             // tell ftclib its inputs  strafeSpeed,forwardSpeed,turn,heading
             drivebase.driveFieldCentric(
                     0.5,
                     0,
                     0,
-                    heading
-
+                    0
             );
-            finger.setPosition(0.0);
 
+            finger.setPosition(0.0);
             telemetry.addData("Path", "Leg 1: %4.1f S Elapsed", runtime.seconds());
             telemetry.update();
         }
         // Step 2:  Stop
-            stop();
+
     }
-}
+}//set finger to 0
