+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AutoPrograms/Auto_ShortLeft.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.AutoPrograms;
 
 import com.arcrobotics.ftclib.drivebase.MecanumDrive;
 import com.arcrobotics.ftclib.hardware.motors.Motor;
@@ -8,8 +8,8 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.IMU;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
-@Autonomous(name = "Long_Left",group = "")
-public class Auto_LongLeft extends LinearOpMode {
+@Autonomous(name = "Short_Left",group = "")
+public class Auto_ShortLeft extends LinearOpMode {
     private final ElapsedTime runtime = new ElapsedTime();
     private Motor leftFrontDrive = null;
     private Motor rightFrontDrive = null;
@@ -47,28 +47,16 @@ public class Auto_LongLeft extends LinearOpMode {
         waitForStart();
         imu.resetYaw();
         // Step through each leg of the path, ensuring that the Auto mode has not been stopped along the way
-        //step 1 : move forward to clear frame
-        // Step 2:  strafe right for X seconds:
+
+        // Step 1:  strafe left for X seconds:
         // Replace with encoder measured distance if needed
 
         runtime.reset(); // reset timer
-        //move forward 2 inches
-        while (opModeIsActive() && (runtime.seconds() < 0.04)){
-            // tell ftclib its inputs  strafeSpeed,forwardSpeed,turn,heading
-            drivebase.driveFieldCentric(
-                    0,
-                    0.5,
-                    0,
-                    0
-            );
-            telemetry.addData("Path", "Leg 1: %4.1f S Elapsed", runtime.seconds());
-            telemetry.update();
-        }
-        while (opModeIsActive() && (runtime.seconds() < 3.0)) {
+        while (opModeIsActive() && (runtime.seconds() < 2.0)) {
             // tell ftclib its inputs  strafeSpeed,forwardSpeed,turn,heading
             drivebase.driveFieldCentric(
                     -0.5,
-                    0.0,
+                    0,
                     0,
                     0
             );
