+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AutoPrograms/Auto_ShortLeftv2.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.AutoPrograms;
 
 import com.arcrobotics.ftclib.drivebase.MecanumDrive;
 import com.arcrobotics.ftclib.hardware.motors.Motor;
@@ -9,30 +9,31 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.IMU;
 import com.qualcomm.robotcore.util.ElapsedTime;
-@Autonomous(name = "New_LongRight_Auto",group = "")
+
+
+@Autonomous (name = "Short_Left2",group = "")
 @Disabled
-public class New_LongRight_Auto extends OpMode {
+
+public class Auto_ShortLeftv2 extends OpMode {
     private final ElapsedTime runtime = new ElapsedTime();
     private Motor leftFrontDrive = null;
     private Motor rightFrontDrive = null;
     private Motor leftBackDrive = null;
     private Motor rightBackDrive = null;
-    private Motor winch = null;
     private Motor ypod = null;
+    private Motor winch = null;
     private Motor arm1 = null;
-    double xdistance = 0.0;
-    double ydistance = 0.0;
-    double winchspeed = 0.25;
-    int autoState = 0;
     MecanumDrive drivebase = null;
     private IMU imu;// BHI260AP imu on this hub
     private boolean test = false;
     private RevHubOrientationOnRobot orientationOnRobot;
+    double winchspeed = .25;
+    double x_distance = 0.0;
+    double y_distance = 0.0;
+    int autoState = 0;
 
     @Override
     public void init() {
-        ypod = new Motor(hardwareMap, "y_encoder");
-        winch = new Motor(hardwareMap, "winch");
         // Initialize the drive system variables.
         leftFrontDrive = new Motor(hardwareMap, "left_front_drive");
         rightFrontDrive = new Motor(hardwareMap, "right_front_drive");
@@ -53,33 +54,42 @@ public class New_LongRight_Auto extends OpMode {
                 leftBackDrive,
                 rightBackDrive
         );
-
+        // use 'fake' motor to get Y encoder values
+        ypod = new Motor(hardwareMap, "y_encoder");
+        // use 'fake' motor to get X encoder values
+        winch = new Motor(hardwareMap, "winch");
     }
 
     @Override
     public void loop() {
+// The odometry pods labelled "winch" and "ypod" will use distance instead of time.
         // tell ftclib its inputs  strafeSpeed,forwardSpeed,turn,heading
-        // drive forward 2 inch,
-        // strafe 102 inch to right
-       switch(autoState){
-           case 0:
-               if (xdistance > 2.0) {
-                   autoState++;
-                   drivebase.stop();
+        //Drive forward 2 inches
+        //The robot strafes 63 inches
+        switch (autoState) {
+            case 0:
+                if (y_distance > 2.0) {
+                    drivebase.stop();
+                    autoState++;
+                }
 
-               }
-               else drivebase.driveFieldCentric(0.0, -0.5,0.0,0.0);
-               break;
-       }
-        drivebase.driveFieldCentric(
-                0,
-                0.5,
-                0,
-                0
-        );
+                else drivebase.driveFieldCentric(0.0, -0.5, 0.0, 0.0);
+                break;
+            case 1:
+                if (x_distance > 63.0) {
+                    drivebase.stop();
+                    autoState++;
+                }
+
+                else drivebase.driveFieldCentric(0.5, -0.0, 0.0, 0.0);
+                break;
+        }
+
+
         telemetry.addData("Path", "Leg 1: %4.1f S Elapsed", runtime.seconds());
         telemetry.update();
         updateOdometry();
+
     }
 
     @Override
@@ -87,17 +97,15 @@ public class New_LongRight_Auto extends OpMode {
         imu.resetYaw();
         runtime.reset();
         arm1.resetEncoder();
-        ypod.setDistancePerPulse((48.0 * Math.PI) / 2000.0);
         winch.setDistancePerPulse((48.0 * Math.PI) / 2000.0);
+        ypod.setDistancePerPulse((48.0 * Math.PI) / 2000.0);
         ypod.resetEncoder();
         winch.resetEncoder();// this motor's encoder is used for Xpod
 
     }
 
     private void updateOdometry() {
-        xdistance = winch.getDistance();
-        ydistance = ypod.getDistance();
+        x_distance = winch.getDistance();
+        y_distance = ypod.getDistance();
     }
 }
-    //set finger to 0
-
