+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/JustTesting/CameraCalibration.java
@@ -31,7 +31,7 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.JustTesting;
 
 import android.util.Size;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
