+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/JustTesting/ServoTest.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.JustTesting;
 
 
 import com.arcrobotics.ftclib.gamepad.GamepadEx;
